% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHubApp.R
\name{AnnotationHubApp}
\alias{AnnotationHubApp}
\title{Initialize the shiny application for AnnotationHub resources}
\usage{
AnnotationHubApp(...)
}
\arguments{
\item{...}{Further arguments to the \code{runApp} function}
}
\description{
The shiny app will allow the user to view a table of \code{AnnotationHub}
resources. The user can select a single resource for download or select
multiple resources to download a \code{data.frame} of metadata that includes
\code{AnnotationHub} identifiers used to download, i.e. in pseudocode:
\code{AH[["Identifier"]]}.
}
\details{
Note. The code here was adapted from \code{interactiveDisplayBase} and
\code{?'display,Hub-method'}.
}
\seealso{
?\code{interactiveDisplayBase::display}, ?\verb{display,Hub-method}
}
