% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationHubApp.R
\name{AnnotationHubApp}
\alias{AnnotationHubApp}
\title{Initialize the shiny application for AnnotationHub resources}
\usage{
AnnotationHubApp(...)
}
\arguments{
\item{...}{Further arguments to the \code{runApp} function}
}
\description{
The shiny app will allow the user to view a table of \code{AnnotationHub}
resources including \code{AnnotationHub} identifiers which are important for
download.
}
